fixed individual ID bug
